OneDrive/Desktop/Emotion_Detection_Project/requirements.txt
streamlit
deepface
opencv-python-headless
numpy
tensorflow
Pillow
